@@ -13,15 +13,14 @@
     "husky": "^4.2.5",
     "lint-staged": "^10.2.13",
     "prettier": "^2.1.0",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.3.0",
-    "react-datepicker": "^3.2.2",
-    "react-dom": "^16.13.1",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.2",
+    "react-datepicker": "^6.9.0",
+    "react-dom": "^18.3.1",
     "react-extra-prop-types": "^0.3.1",
     "react-icons": "^3.11.0",
     "react-player": "^2.6.2",
     "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.1",
     "react-syntax-highlighter": "^15.5.0",
     "react-toastify": "^6.0.8",
@@ -38,7 +37,8 @@
   },
   "devDependencies": {
     "dotenv": "^8.2.0",
-    "prop-types": "^15.7.2"
+    "prop-types": "^15.7.2",
+    "react-router-dom": "^6.23.0"
   },
   "husky": {
     "hooks": {
