@@ -1,13 +1,13 @@
 import React from 'react';
-import { Redirect } from 'react-router-dom';
+import { Navigate } from 'react-router-dom';
 
-const Callback = props => (
-        ({handleAuthentication}) => {
-            if(/access_token|id_token|error/.test(props.location.hash)) {
-                handleAuthentication();
-            }
-            return <Redirect to="/"/>;
-        }
-);
+const Callback =
+  (props) =>
+  ({ handleAuthentication }) => {
+    if (/access_token|id_token|error/.test(props.location.hash)) {
+      handleAuthentication();
+    }
+    return <Navigate to="/" />;
+  };
 
 export default Callback;
